Refactoring
- Move calculations.helpers.roundup to utils and make more generic
- Move lookup_id_to_uuid, etc. to validation services
- Fix some argument names and comments
- Add some new utils
