+++ backend/hitas/utils.py
@@ -1,14 +1,13 @@
 import datetime
 import operator
-from typing import Any, Iterable, Optional, TypeVar
+from decimal import ROUND_HALF_UP, Decimal
+from typing import Any, Iterable, Optional, overload
 from uuid import UUID
 
 from django.db import models
 from django.db.models import Count, Model, OuterRef, Subquery, Value
 from django.db.models.functions import Round
 from django.utils import timezone
-
-from hitas.exceptions import HitasModelNotFound
 
 
 class RoundWithPrecision(Round):
@@ -86,29 +85,6 @@ def valid_uuid(value: str, version: int = 4) -> bool:
         return False
 
 
-TModel = TypeVar("TModel", bound=models.Model)
-
-
-def lookup_id_to_uuid(lookup_id: str, model_class: type[TModel]) -> UUID:
-    try:
-        return UUID(hex=lookup_id)
-    except ValueError as error:
-        raise HitasModelNotFound(model=model_class) from error
-
-
-def lookup_model_by_uuid(lookup_id: str, model_class: type[TModel], **kwargs) -> TModel:
-    uuid = lookup_id_to_uuid(lookup_id, model_class)
-
-    try:
-        return model_class.objects.get(uuid=uuid, **kwargs)
-    except model_class.DoesNotExist as error:
-        raise HitasModelNotFound(model=model_class) from error
-
-
-def lookup_model_id_by_uuid(lookup_id: str, model_class: type[TModel], **kwargs) -> int:
-    return lookup_model_by_uuid(lookup_id, model_class, **kwargs).id
-
-
 def check_for_overlap(range_1: set[int], range_2: Iterable[int]) -> tuple[Optional[int], Optional[int]]:
     overlapping: set[int] = range_1.intersection(range_2)
     if not overlapping:
@@ -125,3 +101,22 @@ def check_for_overlap(range_1: set[int], range_2: Iterable[int]) -> tuple[Option
 
     sorted_shares: list[int] = sorted(overlapping)
     return sorted_shares[0], sorted_shares[-1]
+
+
+@overload
+def roundup(v: Decimal, precision: int = 2) -> Decimal:
+    ...
+
+
+@overload
+def roundup(v: None, precision: int = 2) -> None:
+    ...
+
+
+def roundup(v, precision: int = 2):
+    if v is None:
+        return None
+
+    if precision <= 0:
+        return v.quantize(Decimal("1"), ROUND_HALF_UP)
+    return v.quantize(Decimal("." + "0" * precision), ROUND_HALF_UP)
